requirements.txt
# Core
streamlit>=1.32.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.0

# Data
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# LLM
groq>=0.4.2
openai>=1.12.0

# Vector DB
chromadb>=0.4.22
faiss-cpu>=1.7.4
sentence-transformers>=2.5.1

# Geospatial
geopandas>=0.14.3
folium>=0.16.0
streamlit-folium>=0.18.0
pydeck>=0.8.0
shapely>=2.0.3

# ML
scikit-learn>=1.4.0

# Visualization
plotly>=5.19.0
matplotlib>=3.8.3
seaborn>=0.13.2

# Storage
sqlalchemy>=2.0.28

# Scraping fallback
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Utilities
tenacity>=8.2.3
loguru>=0.7.2
rich>=13.7.1
tqdm>=4.66.2

# Optional API
fastapi>=0.110.0
uvicorn>=0.27.1
